airflow/scripts/requirements.txt
numpy==1.26.2
pandas==2.1.0
yfinance==0.2.28
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
